+++ settings.gradle
@@ -8,4 +8,5 @@ include 'rxjava-core', \
 'rxjava-contrib:rxjava-swing', \
 'rxjava-contrib:rxjava-android', \
 'rxjava-contrib:rxjava-apache-http', \
-'rxjava-contrib:rxjava-string'
+'rxjava-contrib:rxjava-string', \
+'rxjava-contrib:rxjava-async-util'
